1523:tests for Enter handling in text blocks
