+++ components/organisms/editor/tools/utils.ts
@@ -1,10 +1,5 @@
-import { KeyboardEvent } from 'react';
-
-import { IEditorAPI, IEditorContext } from '../context/EditorContext.types';
-import { IBlockData, IBlockNode, TInitialBlockData } from '../instance/Editor.types';
-import { isEmpty } from '../utils/html';
-import { Selection } from '../utils/selection';
-import { IParagraphData } from './paragraph';
+import { IEditorContext } from '../context/EditorContext.types';
+import { IBlockData, IBlockNode } from '../instance/Editor.types';
 
 export const generateBlockNode = (element: HTMLDivElement, block: IBlockData): IBlockNode => {
   // just creating new object from params will not work
@@ -16,135 +11,6 @@ export const generateBlockNode = (element: HTMLDivElement, block: IBlockData): I
   } satisfies IBlockData);
 };
 
-const textNodeNames = 'h1, h2, h3, h4, h5, h6, p';
-
-export const textBlockKeydownHandler = <T extends { text: string }>(
-  event: KeyboardEvent,
-  api: IEditorAPI,
-  element: HTMLElement,
-  block: Pick<IBlockData, 'id' | 'type'>,
-  config?: { shouldMergeOnBackspace?: boolean },
-) => {
-  const { code } = event;
-  const { id: blockId, type: blockType } = block;
-
-  // if user pressed Enter with ctrl or command add new paragraph
-  if (code === 'Enter') {
-    const range = Selection.range;
-    if (!range) return;
-
-    const cursorPosition = range.startOffset;
-    const currentText = element.innerText;
-
-    // Split the text at the cursor position
-    const beforeCursorText = currentText.slice(0, cursorPosition);
-    const afterCursorText = currentText.slice(cursorPosition);
-
-    // if there is no text after the cursor -> only add a paragraph
-    if (!afterCursorText.trim()) {
-      api.addBlock('paragraph', blockId);
-      event.preventDefault();
-      return;
-    }
-
-    // update current block state
-    api.setBlock<IParagraphData>({
-      id: blockId,
-      type: blockType,
-      data: { text: beforeCursorText },
-    });
-
-    // update block content (block will not be rendered when text content is updated)
-    const currentBlockElement = document.getElementById(blockId)!;
-    const textElement = currentBlockElement.querySelector(textNodeNames)!;
-    textElement.innerHTML = beforeCursorText;
-
-    // add a new block
-    api.addBlock('paragraph', blockId, { text: afterCursorText }).then((newBlock) => {
-      const newBlockElement = document.getElementById(newBlock.id);
-      const paragraph = newBlockElement?.querySelector('p');
-      paragraph?.focus();
-    });
-
-    event.preventDefault();
-  }
-
-  // if paragraph is empty when user presses Back key remove this paragraph
-  if (code === 'Backspace' && isEmpty(element)) {
-    api
-      .removeBlock(blockId)
-      .then(({ prevBlocks }) => focusPreviousBlock(prevBlocks, blockId, event));
-    return;
-  }
-
-  if (code === 'Backspace' && config?.shouldMergeOnBackspace) {
-    const selection = Selection.selection;
-    if (!selection) return;
-
-    const range = selection.getRangeAt(0);
-    const offset = range.startOffset;
-
-    if (offset === 0) {
-      api.mergeWithPrevBlock<T>(blockId, (prevBlock, currentBlock) => {
-        prevBlock.data.text += ` ${currentBlock.data.text}`;
-
-        return prevBlock.data;
-      });
-      return;
-    }
-  }
-};
-
 export function getCurrentBlock<T>({ data, hoveredBlock }: IEditorContext<T>) {
   return data.find((b) => hoveredBlock?.id === b.id);
-}
-
-function getLastTextElement(element: HTMLElement) {
-  const textElements = element?.querySelectorAll(textNodeNames) as NodeListOf<
-    HTMLParagraphElement | HTMLHeadingElement
-  >;
-
-  if (!textElements || textElements.length < 1) return null;
-
-  return textElements[textElements.length - 1];
-}
-
-function focusPreviousBlock(
-  prevBlocks: IBlockData<TInitialBlockData>[],
-  blockId: IBlockData['id'],
-  event: KeyboardEvent,
-) {
-  const prevBlock = prevBlocks.find((_, index) => {
-    const nextBlock = prevBlocks[index + 1];
-    // if next block is the deleted block -> current block is previous block
-    if (nextBlock?.id === blockId) return true;
-
-    return false;
-  });
-
-  if (!prevBlock) return;
-
-  const prevBlockElement = document.getElementById(prevBlock?.id);
-
-  const lastTextElement = getLastTextElement(prevBlockElement!);
-  if (!lastTextElement) return;
-  lastTextElement.focus();
-
-  // Create a range (a span of content) and a selection (current user selection)
-  const range = document.createRange();
-  const selection = window.getSelection();
-
-  if (!selection) return;
-
-  // Move the range to the end of the content inside the element
-  range.selectNodeContents(lastTextElement);
-  range.collapse(false);
-
-  // Remove any current selections
-  selection.removeAllRanges();
-
-  // Add the new range to the selection
-  selection.addRange(range);
-
-  event.preventDefault();
 }
